Downloads/Bot/requirements.txt
python-binance==1.0.19
tensorflow==2.15.0
pandas==2.1.4
numpy==1.26.2
ta==0.11.0
python-dotenv==1.0.0